more

+++ packages/hebao_v1/contracts/stores/Data.sol
@@ -5,10 +5,11 @@ pragma solidity ^0.7.0;
 
 library Data
 {
+    // This struct should take less than 32 bytes.
     struct Guardian
     {
         address addr;
-        uint64  group;
+        uint8   group;
         int64   timestamp; // If < 0, `timestamp.abs` is the validUntil timestamp;
                            // If > 0, `timestamp` is the validSince timestamp;
     }
